Refactor day summary builder into pipeline stages
